# vecta/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Vecta Platform API
  version: 1.0.0
  description: |
    Life-as-a-Service for F-1 international students.
    
    **Authentication:** All `/api/v1/*` endpoints (except webhooks) require
    `Authorization: Bearer <VectaIDToken>` (RS256 JWT).
    
    **F-1 Compliance:** `POST /api/v1/mobility/rides/accept` and
    `POST /api/v1/mobility/driver/go-online` always return `403 F1_VISA_COMPLIANCE_VIOLATION`
    regardless of token — these routes are architectural dead ends.
  contact:
    email: api@vecta.io
  license:
    name: Proprietary
    url: https://vecta.io/terms
servers:
  - url: https://api.vecta.io
    description: Production
  - url: https://api.staging.vecta.io
    description: Staging
  - url: http://localhost:4000
    description: Local development

# ---------------------------------------------------------------------------
# Security schemes
# ---------------------------------------------------------------------------
components:
  securitySchemes:
    VectaIDToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 JWT issued by Vecta identity service. Contains zero PII.
  schemas:
...
    TokenRecord:
      type: object
      properties:
        jti:       { type: string, description: Masked JTI (first 8 chars + …) }
        createdAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        used:      { type: boolean }
        usedAt:    { type: string, format: date-time, nullable: true }
    EarningsSummary:
      type: object
      properties:
        taxYear:           { type: integer }
        ytdRentalIncome:   { type: number, description: USD }
        rideCount:         { type: integer }
        activeSince:       { type: string, format: date-time, nullable: true }
        taxClassification: { type: string, example: 'Schedule E — Passive Rental Income' }
        form1099Type:      { type: string, example: '1099-MISC Box 1: Rents' }
...
paths:
...
  /api/v1/identity/tokens:
    get:
      summary: List active sharing tokens for the authenticated student
      security: [{ VectaIDToken: [] }]
      tags: [Identity]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  tokens:
                    type: array
                    items: { $ref: '#/components/schemas/TokenRecord' }
  /api/v1/identity/tokens/{jti}/revoke:
    delete:
      summary: Revoke a sharing token (student withdraws consent)
      security: [{ VectaIDToken: [] }]
      tags: [Identity]
      parameters:
        - name: jti
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  revoked: { type: boolean }
        '404':
          description: Token not found or already used
...
  /api/v1/mobility/earnings:
    get:
      summary: YTD Schedule E earnings summary
      security: [{ VectaIDToken: [] }]
      tags: [Mobility]
      parameters:
        - name: year
          in: query
          schema: { type: integer }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EarningsSummary' }
...
  /api/v1/mobility/audit/chain:
    get:
      summary: Export immutable flight recorder audit chain (USCIS/IRS)
      security: [{ VectaIDToken: [] }]
      tags: [Mobility]
      parameters:
        - name: year
          in: query
          schema: { type: integer }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  chainIntegrity: { type: string, enum: [VERIFIED, BROKEN] }
                  entries:        { type: array, items: { type: object } }
                  exportedAt:     { type: string, format: date-time }

  # --- Landlord ---
  /api/v1/landlord/register:
    post:
      summary: Register landlord account + send magic link
      tags: [Landlord]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:       { type: string, format: email }
                fullName:    { type: string }
                companyName: { type: string }
      responses:
        '201':
          description: Always 201 (email enumeration prevention)
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
  /api/v1/landlord/verify-email:
    post:
      summary: Verify landlord email via magic link token
      tags: [Landlord]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:   { type: boolean }
                  tier: { type: string, enum: [VERIFIED, TRUSTED] }
        '401':
          description: Token invalid, expired, or already used

  # --- Webhooks (public routes — HMAC verified) ---
...